Refactor Navigation rules to support N-Way surveys
