+++ backbone/src/main/java/org/researchstack/backbone/model/navigable/LessThanOrEqualsNavigationClause.java
@@ -3,11 +3,7 @@ package org.researchstack.backbone.model.navigable;
 import org.researchstack.backbone.answerformat.AnswerFormat;
 import org.researchstack.backbone.result.StepResult;
 
-/**
- * Created by mauriciosouto on 10/10/17.
- */
-
-public class LessThanOrEqualsNavigationClause extends StepNavigationClause {
+public class LessThanOrEqualsNavigationClause extends StepNavigationClauseRule {
 
     public LessThanOrEqualsNavigationClause(String sourceStepIdentifier, Object value, AnswerFormat.Type ruleType, RuleClauseOperand operand) {
         super(sourceStepIdentifier, value, ruleType, operand);
