+++ backbone/src/main/java/org/researchstack/backbone/model/navigable/NotEmptyNavigationClause.java
@@ -3,11 +3,8 @@ package org.researchstack.backbone.model.navigable;
 import org.researchstack.backbone.answerformat.AnswerFormat;
 import org.researchstack.backbone.result.StepResult;
 
-/**
- * Created by mauriciosouto on 10/10/17.
- */
 
-public class NotEmptyNavigationClause extends StepNavigationClause {
+public class NotEmptyNavigationClause extends StepNavigationClauseRule {
 
     public NotEmptyNavigationClause(String sourceStepIdentifier, Object value, AnswerFormat.Type ruleType, RuleClauseOperand operand) {
         super(sourceStepIdentifier, value, ruleType, operand);
