+++ backbone/src/main/java/org/researchstack/backbone/model/navigable/RuleClauseOperand.java
@@ -1,9 +1,5 @@
 package org.researchstack.backbone.model.navigable;
 
-/**
- * Created by mauriciosouto on 10/10/17.
- */
-
 public enum RuleClauseOperand {
     NONE,
     AND,
