+++ skin/build.gradle
@@ -5,7 +5,7 @@ apply plugin: 'com.neenbedankt.android-apt'
 apply plugin: 'com.jfrog.bintray'
 apply plugin: 'maven-publish'
 
-version = '2.0.97'
+version = '2.0.99'
 
 android {
     compileSdkVersion 25
@@ -18,7 +18,7 @@ android {
     defaultConfig {
         minSdkVersion 16
         targetSdkVersion 25
-        versionCode 104
+        versionCode 106
         versionName version
     }
     buildTypes {
